Refactor_Streaming/Fases/fase1/requirements/base.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
drf-spectacular>=0.26.0
django-cors-headers>=4.0.0
python-decouple>=3.8
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
django-extensions>=3.2.0